Tilemap+physics (#17)
* cleanup

* cleanup

* initial commit

* version update

* fixed collider not working

Replaced hit.normal values with checks for if they are positive or negative. To make sure hitPosition never gets 0 as a value when normal.x or normal.y is 0.

* added some extra tags for no reason

* trimmed the sprites and added physics outline

* internal changes and tags

* fixed collision (for real this time) and code cleanup

* Updated move and coins

Added coin collection and fixed jumping

* add physics shape

* added player tag

* meta files updated

* replaced tag Floor with Ground

* fixed reference to maincamera, added coins, money text

* Add help popup and other fixes (#11)

* prefab for help popup text

* script to manage popup text

Generate popup text based on name of the object the script was attached to

* remove script from prefab

Script needs to be attached to the physics/collider substances/objects, not to this

* ignore another custom folder

* updated help text

Removed default message and replace slime with slippery

* fixed popup generation

Generate popup on trigger and reset after 2 seconds. Queue popup when multiple triggers are called concurrently.

# Conflicts:
#	Assets/Scenes/Level1.unity

* remove visual scripting

* fix start button

* recreate help popup prefab in a canvas

Added the corner buttons and the popup into the canvas prefab

* add todo

Gameobject.Find is a inefficient method in this context

* update name of canvas containing the popup

* initialize reference to self in the script

Public variable to set the reference to tmpro gameobject is unnecessary

* fix pause button

* various minor tweaks and fixes

* updated settings

* added eventsystem

* pause button and fullscreen button fixed (on Level1)

* delete maingame

* delete sample scene

* added pause menu camera and dummy menu canvas

* remove pause menu dummies

Preparing to switch to urp to get overlay cameras

---------

Co-authored-by: arhart1 <113941310+arhart1@users.noreply.github.com>